requirements.txt
python-dotenv==1.0.1
openai==1.51.2
faster-whisper==1.0.3
sounddevice==0.4.7
numpy==2.1.1
gtts==2.5.3
soundfile==0.12.1
httpx==0.27.2
httpcore==1.0.5
pygame==2.5.2
pjsua2==2.13.1
threading-timer==0.1.0

